Rewrite of heart rate monitor application (intended for further development)
Author: John C. Peterson <jcp-sd@users.noreply.github.com>
Signed-off-by: John C. Peterson <jcp-sd@users.noreply.github.com>

This is a comprehensive rewrite of the heart rate monitoring application
and associated support files. It is designed with further research and
prototyping of more sophisticated heart rate estimation algorithms in
mind.

It fixes the known sampling rate issues of the HRS3300 and includes a new,
improved scheme for polling the HRS3300 at more precise intervals. It
includes a new IIR filter class that supports filters with an arbitrary
number of poles and zeros. See the ppg-filtering directory for the
scripts to pre-calculate the filter coefficients used by default.

The heart rate estimation algorithms operate on a buffer of filtered
samples of configurable length (10 seconds is the default). Two algorithms
are presently implemented; one based on a peak finder, and the other based
on the sample autocorrelation function. To avoid dropping PPG samples from
one massive computation when the buffer is full, much of the computational
effort has been "spread out" over subticks, and performed after the PPG
sample has been filtered.

The new app still drops or skips over samples every so often. Getting to
that point took a lot of effort! It was discovered that writing text to
the display is surprising quite slow. Updates to the display had to be
"spread out" over multiple subticks.

The new app takes up more room that the original app, and some apps had "to
go" from the binary image to make room for it. See wasp/apps/software.py
and also wasp/boards/manifest_240x240.py for the ones I decided to delete
from the image.

Runtime memory is also a bit tight, and you might need to disable some
apps (from the "Software" app) to get the new HRM app to run.