Updates to PPG filter coefficients
Author: John C. Peterson <jcp-sd@users.noreply.github.com>
Signed-off-by: John C. Peterson <jcp-sd@users.noreply.github.com>

Some tweaks to the PPG filter coefficients. The passband of the HP filter
now starts around 40 bpm. The previous value of about 30 bpm was much too
low as it sometimes passed harmonics that appeared to come from (heavy)
breathing. The level of the stopband of the LP filter was lowered to 30
dB (from 20 dB). The performance is significantly improved. It is much
more robust in the presence of minor motion.